Use docker-compose library instead of fork subprocesses
This patch introduces the usage of docker-compose library, to perform
the `up` and `down` commands. It avoids to fork subprocesses and the
pytest fixture can use the compose.Project object to deal with the
docker-compose.yml configuration and your containers.

In order to replace subprocess calls by the usage of the docker-compose
library it is required to update the current tests. This patch
replaces the mock of subprocess by the direct call of `docker-compose`.
I hope that github CI supports docker to support these tests.

Tests about DockerComposeExecutor are removed, the next patch will
rework the DockerComposeExecutor and the Services classes.